chore: modify the location of the find_key_index function.
