+++ src/pipeline/src/etl/processor/dissect.rs
@@ -19,10 +19,11 @@ use common_telemetry::warn;
 use itertools::Itertools;
 
 use crate::etl::field::{Fields, InputFieldInfo, OneInputMultiOutputField};
+use crate::etl::find_key_index;
 use crate::etl::processor::{
-    find_key_index, yaml_bool, yaml_new_field, yaml_new_fields, yaml_parse_string,
-    yaml_parse_strings, yaml_string, Processor, ProcessorBuilder, ProcessorKind, FIELDS_NAME,
-    FIELD_NAME, IGNORE_MISSING_NAME, PATTERNS_NAME, PATTERN_NAME,
+    yaml_bool, yaml_new_field, yaml_new_fields, yaml_parse_string, yaml_parse_strings, yaml_string,
+    Processor, ProcessorBuilder, ProcessorKind, FIELDS_NAME, FIELD_NAME, IGNORE_MISSING_NAME,
+    PATTERNS_NAME, PATTERN_NAME,
 };
 use crate::etl::value::Value;
